fix(realtime): rename Plivo file
