+++ packages/agents-extensions/src/index.ts
@@ -1,3 +1,3 @@
 export * from './TwilioRealtimeTransport';
 export * from './aiSdk';
-export * from './Plivo';
+export * from './PlivoRealtimeTransport';
